#2 -- Initial application skeleton
This is effectively the "hello world" stage of the project, with essentially nothing but generated code. All it does is display "Hello Play!". But it's a start.

This is based on the play-scala-seed giter8 template. The only change is a tweak to the routes file, to fix a new Scala 2.13.3 warning.