Introduce the LoginController
This it the camel's nose in the tent, that forces us to get serious about the server structure, so lots of code here.

Introduces a reasonably best-practices listing of dependencies, so that it's easy to see what we're using.

Pulls in Macwire for dependency injection, and rewrites everything to make use of that.