#63 - read in the KonOpas-formatted data file
This brings us to the point where we have a unit test that contains the KO data from 2014, and shows that we can read that into Scala data structures. That's not by any means done, but it's the starting point for enriching that data.

Includes another test copy of the KonOpas data, where the test wants it. There may be yet another copy soon, but whatever - we'll live.

The ScheduleService isn't actually hooked into the system yet, but we at least have some evidence that it is, in and of itself, working.

Note that this comments out the HomeControllerSpec, which will be deleted in due course.