Provide the Schedule for the front end
Adds a new `GET /api/schedule` entry point. Under normal circumstances, this returns OK with:
```
{"program":"... the program items","people":"... the people"}
```
`program` and `people` should be in more or less exactly the format from KonOpas -- the backend reads that in, parses it, will eventually enhance it, and then re-serializes the JSON.

For now, this is dummy data (the same 2014 data set as usual), but this entry point will eventually serve the real schedule from Zambia.

Along the way, let's start being good about returning the right Content-Type.