Get a production build working
This enables the `dist` command in sbt. This command constructs a zip file in the backend target/universal directory -- we should unzip that file into the actual run directory and run the bash script contained therein in order to run the server for real.

This still needs to be actually tried out on the server, but the results look right.

Before building, the secrets.conf.template file needs to be copied to secrets.conf, and the bits filled in. Going forward, we will need to keep the real secrets.conf in line with changes in secrets.conf.template.

This also contains some checkpoint code towards setting up the database; it's a start, but a good ways from doing anything real yet.