Introduce database evolutions
In the Play world, the right way to deal with the DB is to use Evolutions, which provides schema evolution capability. When you boot the server, it looks at the defined evolutions, and adds any that aren't yet performed.

We will probably be a bit loose-and-sloppy about this during early development, when it's still easy to blow away your DB and start again. But as we get down to the wire, this will be helpful for keeping the DB intact as things change at the last minute.

Note that, at this point, you need to have a Postgres instance running, and set the connection URL to it in secrets.conf.