Fetch the Schedule from pseudo-Zambia
We don't yet have a real Zambia instance to fetch the Schedule from. Worse, we can't really use the real data for testing, since we need to test events that are happening "now".

So this introduces the beginnings of a loopback FakeZambiaController -- a fake of Zambia that provides a fake Schedule. That allows us to make the Schedule-handling logic, including the fetching, real -- we're just pointing to a fake entry point for now. That way, we can really start fleshing things out.

So we now do a refresh from "Zambia" on a regular clock tick -- initially hard-coded to five minutes, but that will become configurable.

Since the Schedule is now always cached, and initialized to an empty default at startup, the getSchedule() pathway is now synchronous. This may change, but we'll see.

Fixed a bug in the Play Evolutions, which have a problem with semicolons.