Add "Infinite" Image Carousel.
This carousel works by having an arbitrarily large list of images. It
has a buffer of 3 images and creates a carousel with those three
images. Based on the current image and the current frame of the
carousel, it updates the buffer so that the images on either side of
the current image will be correct.

This allows us to have an arbitrarily large set of images while only
having 3 sets of DOM built at any time. It also only loads the images
when they're the next one to be shown. We can update the size of the
buffer if we want to preload more images. We could even have more
images preloaded ahead than behind.

We can also have a trigger when we're getting close to the currently
loaded set of images to ask for more images. I'm imagining a backend
that returns images in chunks. When we're within 5 images or so of the
end of the current set of images, we request a new chunk.  We can have
the frontend pick a seed and have the backend run a permutations
algorithm based on that seed, thus guaranteeing that a user won't see
the same image until they've seen all of them, but they'll be in a
different order each time.

The UI has a gradient overlay at the bottom so that the caption will
be readable, even if the image is white there. It has a pause button
in the lower right corner to allow the user to pause the images if
they wish to. The forward and back buttons will continue to work.

The component claims to support swipes on a touch device, but I
haven't been able to get it to work on chrome emulation.