Rework the code to expect the schedule as JSON instead of JSONP
Gail got the schedule exporter stood up in Zambia, and changed its export format since we prefer JSON. So this adjusts at our end.

For the moment, I'm still using the 2014 JSONP schedule for testing; that will likely get changed sometime in the near future, as the 2021 Zambia schedule gets into a state that works for us.