Adjust the dates in the Schedule to be "now"
For the moment, we're still using the 2014 Zambia data, but we're massaging it so that it always produces a Schedule that consists of yesterday/today/tomorrow/day-after -- all of the dates are pushed forward to be "now". This will permit us to test what things look like when the site is live day-of.

(Obviously this will eventually have to go behind a switch of some sort, and we'll probably switch over to the real A'21 data sometime soon, but this is a step in the right direction.)