Add a config flag so that you can set yourself as an Admin in your dev environment
If you would like to be able to use the Admin UI in your development environment, add your CM username to `arisia.dev.admins` in `secrets.conf`. For example, in my `secrets.conf`, I would have:
```
arisia.dev.admins = ["jducoeur"]
```
Along with this, did a bit of refactoring to keep this sort of config mungery in better-contained places, so the controller doesn't need to worry about it.