Beginnings of Zoom automation
This doesn't do much yet, but we now have a test entry point that fetches the user info for our account. (Which we don't actually care about, but it's the simplest entry point.) This demonstrates that the underlying JWT-generation is working, which was the biggest technical worry.

Pulls in what seems to be the best-reputed JWT library for Scala. Note that that assumes you are doing this for *user*-level JWTs, so it drives itself off of the standard Play config values. We want very different config for Zoom integration (among other things, we want very short expirations), so we construct a separate JWT-specific config for that purpose.