Actually start Zoom meetings
This is only the beginning of the Zoom-meeting functionality, but it's the critical nugget: we can now start meetings from the backend.

This adds an Admin endpoint with (intentionally) no UI, to create ad-hoc meetings, for testing all of this. This can basically only be called by Admins from Swagger, and is hard to use, so I don't expect anybody to abuse it.

Being able to do this is *enormously* informative, and may require a tweak or two to our plans, but it's mostly working as expected.