Make the schedule properly cacheable on the frontend
This adds an `ETag` header from the getSchedule() endpoint -- an MD5 hash of the current schedule. The frontend should return that in an `If-None-Match` header, and the backend will simply return 304 (Not Modified) in the usual case that nothing has changed.

Note that the HTTP standard expects the hash to be in double quotes in both the `ETag` and `If-None-Match` headers: the backend is expecting that.

This also tweaks things to cache the JSON-serialized form of the Schedule, so we don't have to recompute that on every fetch.