Fix logging
Figured out that the reason we weren't writing anything to `application.log` was that the line to invoke that was commented out in the original template we started from. Uncommenting that sends stuff to the log as expected.

Also, the original template was writing all of the console coloring to the log file, which is silly: it just causes problems there. So reformatted the output-file stuff to read better from the file.