Make the header stretch all the way across the page.
This required changing where our top level class="container" lives to
not include the navigation component. Instead the navigation component
is set to 100% width and sticky-top and has the background color. An
extra div is added between the navigation component and the actual
navbar. This div has the "container" type so that the navigation
elements will stay the correct width.