Process schedule data into a form that is more usable.
It creates new classes for ScheduleEvent and SchedulePerson.

In particular, this creates links between events and people.

It also parses the timeStamp string and uses that to create a Date
object for events.

It creates a new StructuredEvents type that is list based instead of
object based.  This object is created by sorting the events and then
partitioning them based on Date.

It changes date filters to be based on Date objects instead of
comparing to date string. This will allow us to search by current time
as well as day of the week. It also means date filters will be based
on current time zone.