Checkpoint: Lots of schedule and queue changes
Several new fields added to ProgramItem, all in the name of Prep sessions: these keep track of which real sessions they are prep for, and they are responsible for tracking the start and stop times of the Zoom meetings. These prep sessions are what we actually use in the Queue Service for starting and stopping meetings.

Significant rewrites to the ScheduleService, to synthesize the Prep Sessions.

The Queue Service now registers itself with a frequent timer.

Created the zoom_rooms and active_program_items tables, which we will begin to use shortly.