Add handling of time zones.
Shows EST times as well when in other time zones.

Adjusts filter days shown based on time zone. Allows searching for any
events from 5:30PM Friday 2:30PM Monday. There are some ways we could
make this see the events in the schedule and base the dates on that,
but this is esaier for now.