Finish the start-meeting workflow
This now actually starts the meeting at the appropriate time.

Note that the signature for starting meetings has changed, because we now have potentially many Zoom "users" owning them.

Introduces the Running Items Queue, and the Currently Running Items Map.

Slims down the ZoomMeeting data structure to only the fields we really care about, to make it easier to save and restore it.

Tweaks 5.sql to add the end time for the active meeting. This is against the rules (since 5.sql was already merged), but I doubt anybody has any data in that table, so I'm allowing myself to cheat.