Make initialization less ad-hoc
This formalizes a simplistic LifecycleService, so that stuff that needs to happen at initialization does so in a more orderly way. Previously, everything was happening in constructors, and while nothing was *breaking* per se, it was throwing enough errors sometimes to be pretty scary. This should, in theory, work better.