Make event timestrings auto update when 12 hour/24 hour changes.
This means we had to remove the cache in events that holds the time
string so that it can update when the time changes. We may add a
global cache from time to string to speed things up. Having it be a
global cache would allow us to clear it as is appropriate.