First draft of the backend for Ducks
This isn't complete yet, mostly because I'm not sure of the precise desired semantics of the requesting_url. Does this mean the Referrer header of the incoming request? Is it mandatory, or only for some ducks?

Similarly, is the hint text mandatory or optional?

Currently I have both hint and requesting_url as optional, but that's purely a guess -- let's get this right before we merge this.