Optimize id filters in scheduleService.
Before this change, id filters walked the entire Event db comparing
each item to each of the ids in the filter.

This changes it so that we keep a map from id to Event (which we were
building anyway and then throwing away), and then takes the list of
ids in the filter and just maps each of them. It then processes the
rest of the filters based on that list.