Adjusted use of requesting_url.
Having slept on it, and realized how requesting_url is probably intended to be used, adjusted everything accordingly:

* It is no longer serialized, so that it can't be trivially scraped.
* We now compare the incoming Referer header against the requesting_url, to validate that the assignment is legit.
* It is no longer considered optional. (I'm leaving "hint" optional, mostly to illustrate that pattern.)

It's still not completely done -- in particular, we don't have the Edit or Delete parts of the CRUD UI, because I haven't had time to build them. Drop me a line if you want guidance on what that should look like. (Most of the pieces exist; I just haven't had time to establish the pattern yet.)

And of course, the API side is largely untested, so be aware that there might be bugs.