Update schedule/people page.
This replaces the list of links on the people page with a set of
search terms and an embedded list of person objects.

PersonComponent no longer reads its data from ActivatedRouter and
instead takes a person as an input. It also acts as an expando, with
the default state being just the participant's name.

schedule/people/<string> is now a prefix search, instead of being an
individual lookup. schedule/people/goh is a special case that displays
the three guests of honor.