Draft of the CM database fetch
Lots of code involved here, including a fairly crazy MySQL query.

All of this is completely untested (it can only be tested on the production side, for security reasons), but the code is getting there.

Many, many new config options, some of which are probably constants and some of which aren't.

Includes a simplistic "test mode" for when CM integration is disabled (anywhere other than the production site), with a few special-case logins so we can test the various error cases. (No account, no current membership, missing the current CoC.) Also includes a configurable ability to specify the desired badge ID in secrets.conf, so you can test *as* a particular user. (Which we will need for developing the Program Participant bits.)