Extract image carousel into a component.
This component has one input. images is a list of carousel images to
display. It doesn't currently support expanding this list of images on
the fly, but that could be added later.

This is just a refactor. The only behavior changes should be an extra
div in the output, but no effect on the user experience.