Add Door dragon to map links.
Adds a directive for links that require the user to be logged in. It
pops up a door dragon if the user isn't logged in. Requires a
doorDragonStyle of either programming or social.

The Door Dragon also supports styles of 'discord' and 'reset', but the
links for those still go to the login screen, so they're not ready yet.