Add a door dragon for password resets.
Changed it so that the door dragon simply reports a click on the OK
button and lets the caller take action. Changes the login-guarded-link
directive to take that action.

Implemented the call to door dragon in the login page when the user
clicks "I forgot my password".

Fixed some typos in the discord and reset styles for door
dragon.