Add backend for Settings
This is very simple, and very untested, but straightforward enough that it is fairly likely to work.

`GET /api/settings` returns a dictionary of key/value pairs for this user.
`POST /api/settings` takes a dictionary of key/value pairs, and inserts/updates them in this user's settings.
`DELETE /api/settings/{key}` removes the specified key from the user's settings.

Also: fixes a bug with the DuckService, which wasn't preloading its cache, and therefore was spuriously returning an empty set.

Also: a bit of refactoring to make it easier to declare functions that only work for logged-in users.