Beginnings of the Schedule Test system
This introduces a new Admin UI, that allows you to input fake program items. These exist only in-memory for the duration of this server run, and get merged into the real schedule. The point of this is to allow us to test actually *using* the schedule, with program items coming up in the near future, to ensure that things actually work as intended.

Also includes the very beginnings of backend-side Discord integration: so far, just a proof of concept that we can fetch the members of the Arisia Discord server. This will form part of the ArisiaBot, sharing the responsibilities with the Python Lambda bits.