Add more-nuanced responses from login
The `/api/login` entry point now returns a subtler response when something isn't right.

In the success case, it's the same as it has been -- 200, with the user info.

In the failure case, it returns 401 Unauthorized, with the error data structure:
```
{"success":"false", "message":"--ENUM--"}
```
where `--ENUM--` is one of the following:
* `nologin`: this username/password combo does not match a proper, well-formed CM account. The user should go to Registration and start from the beginning.
* `nomembership`: this user does not yet have a membership for Arisia'21. They should go to CM and buy one.
* `nococ`: this user has not signed the current version of the Code of Conduct. They should go to CM and accept that.
* `notyet`: the site is not yet open to the general public.

These error messages need wordsmithing, but those are the four known failure modes.

This commit also includes some progress towards connecting members to Discord, but there's a fair ways to go on that yet.