Fix login bugs.
The first bug was that we never subscribed to the "me" call in
account.service.ts. Switching from a tap to a subscribe fixed
that. The other calls currently have subscribers, but I'm going to
learn how to multicast Observables so that we can trigger the
login/logout even if the caller doesn't subscribe.

The second bug was that user$ and loggedIn$ both returned a value
before we got the response from me. This meant that we redirected to
the login page before we were able to check if we were already logged
in.