Made active room info update live.
The repeat system summarized:
defer(() => of(generateTwoEvents)).pipe(
  map(twoEvents => runningEvents),
  mergeMap(runningEvents => concat(of(runningEvents), defer(() => timer(generateDelayTime).pipe(ignoreElements())))),
  repeat();
);

So, it's an Observable that returns the next two events (but does it
deferred so that it will generate it based on the time when subscribed
instead of when created).

Then that is mapped to our RunningEvents structure.

Then we chain together two new Observables. The first simply emits the
current value. The second uses that value to calculate the amount of
time before we will need to recalculate and then runs a timer, but
ignores the output event (and only pays attention to the complete
event).

If there are no more events, it will wait for Infinity ms.

And then we repeat that loop forever.

And everything is done in rxjs, so unsubscribing will do the right
thing.