Actually create the logged-in user record
Since we need to start recording the Discord info, by user, in the DB, this crosses the line where we need to start actually putting user info into the DB in the first place.

Note that folks who are already in the system will need to logout/login before they are able to connect to Discord properly, since we will be depending on this.