Cache the member list
Fetching the members from Discord is expensive, so even if the cache is only sometimes going to work (because this will usually be called shortly after somebody joins the server), that's still a worthwhile saving.

Also deals a bit more gracefully if we get an error from Discord, and gives us some hope of being able to diagnose it.