Assisted Discord setup API
This provides two new entry points:

* `GET /api/discord/getAssistSecret` returns a secret that a user can use in order to seek help getting set up with Discord.
* `POST /api/discord/assistAddArisian` allows Help Desk to provide assistance.

The latter entry point requires a `X-Shared-Secret` header, since it is intended solely for the use of Lambda. (Gail and Justin will agree on a secret for production, which goes into secrets.conf.)

It takes the following data structure:
```
{ "badgeNumber": "1111", "secret": "xxxx", "discordId": "30940493039303930" }
```
The secret comes from the other entry point, and validates the user. The Discord ID comes from the slash comment that Help Desk uses.

Includes an enhancement to store the Badge Name in the DB, which had previously been overlooked.