Move creation of CreatorPageComponent.linksToDisplay to ngOnChanges
This was a function that was called from the html
before. Unfortunately, calling a function from the html means it gets
called every time angular wonders if something needs to be changed and
that happens every time we update the schedule every ten seconds.

We then return a newly created list which compares as not equal to the
old list, so not only does it create a new list, but it also rebuilds
all the html causing the browser to relayout.

Putting this in ngOnChanges means that it will always match the
creator being passed in, but won't run if it doesn't change.