Provide config for the frontend
This provides a new entry point:
```
GET /api/config/:name
```
Assuming that that value is in the backend config, it will return 200 with the value as the body. If the name is not known in the backend config, it will return 404.

In the backend config, the subtree `arisia.frontend` has been set aside for this purpose -- the frontend team are encouraged to put their values in there. The real values should go into secrets.conf; I recommend documenting those entries in secrets.conf.template and maybe application.conf, but that is not required.

So if, say, there is a backend config entry named `arisia.frontend.thingy`, you can call `GET /api/config/thingy` to fetch it.