Redo event item layout.
Moves star to the left of all the content using flex.

Adds an expando to the right end of the content.

Moves the Zoom Link to the same line as the time. (Not sure how the
host zoom links look.)

Moved the <li> into the <app-item> so that we can make the clickable area match the whole row.