Add a simplistic Admin Audit Trail
It ain't fancy, but it at least shows us any time an Admin is poking around, and what they are doing.

Coded such that you *have* to have auditing if you are following the standard Admin UI controller patterns, to prevent accidentally forgetting an entry point. This means that even looking at Admin pages will be audited -- not a bad thing on general principles.