Backend hooks for working with "manual" meetings
(Draft: there is time to change this, but let's be concrete about it.)

We have a lot of Zoom spaces that are *not* scheduled (eg, Gaming, Social, Youth), and so don't fit into the framework so far. These meetings are owned by their respective divisions, who will be empowered to open and close the rooms when they see fit. (We will use the "personal meeting" attached to each account, because that is by far the easiest way to deal with it.)

So this adds two new entry points to allow the frontend to allow users into those rooms when appropriate:
```
GET /api/room/open/:name
GET /api/room/enter/:name
```
In both cases, we will agree on the value of `name` -- probably something like `gaming`, `social`, etc. (The backend values are configured in the Admin UI, so the UI can set what it likes.)

The `open` entry point returns:
```
{"running":true}
or
{"running":false}
```
To indicate whether this space is currently open; if not, it should tell the user in some fashion that the room is currently closed.

Assuming the user is logged in, and the room is open, the frontend should open a new tab calling the `enter` entry point, and it will redirect to the Zoom meeting as usual.