Add Discord Sync
This is the third backend story for Discord. Adds a new entry point:
```
PUT /api/discord/sync/:id
```
Note that the `X-Shared-Secret` header must be set: this is intended only for use from Lambda.

The `:id` should be the Discord ID of a member who has already connected from Virtual Arisia to Discord. It will resync their info -- in practice for now, it resyncs their Badge Name.

Not well-tested, because it is unfortunately difficult to use my Discord account for testing. (Since I'm a Site Admin, bot changes to me tend to be rejected by Discord.)