Lift upload limits
We hit a weird error on production when trying to upload a 145k file. It looks like this *could* happen from Play (although it hasn't been doing so in Dev). This adds a couple of config values that should raise the limit to 10MB, if that's where the problem is coming from.