Provide a way to restart meetings
There is a failure mode we have to deal with: the Host does the wrong thing, and shuts down a program item prematurely. At that point, the Zoom meeting is dead -- everyone is kicked out, and nobody (including the Host) can re-enter.

So this adds a new Admin UI page for this emergency. Give it the item number for a program item in this state, and it will properly close the original version, start up a new meeting, and rewire everything. At that point, the Host can re-enter, and then the participants.

Fixes #265