Create Immutable Duck Structures
Reworks Ducks to do most of the work in the service.

Added DuckState that represents a Duck and whether it's hidden.

Made getDuckStates take a Set<number> full of ids to return. If
includeHidden is set to true, all ducks will be included, but only the
ducks in the set will be fully shown.