Get Webinars working
While Zoom's API is mostly nice, it does have a deep stupidity that the entry points for Meetings and Webinars are completely unrelated, even though they do almost exactly the same things.

So this propagates through the concept of whether a given program item is a Webinar or not, and calls the appropriate entry points depending on the answer.