Get Manual Meeting Start URLs
We need to be able to get the magic Start URLs for the manually-managed meetings, so that people other than the account owner can start them. This provides an admin interface that, given the standardized name for the meeting (eg, "social"), hands you back the Start URL, which can be used immediately or shared around.