Separate ETags for Hosts from those for everyone else
Fixes a problem Chris pointed out several days ago, which I failed to find time to investigate. This should solve the problem when people are added/removed as Tech during the con -- it produces a different ETag in those situations.

(I believe it's a non-issue when someone's Program Participant status changes: that implies a change to the Schedule, so the ETags will be invalidated anyway.)