Redirect FastTrack panels into the Youth Services Zoom
Fast Track is a funny case: unlike most of our programming, with individually-created meetings for everything, it's just a breakout room in Youth Services.

So this little hack intercepts all attempts to enter Fast Track panels, and instead routes them to the Youth Services Zoom. Not the prettiest thing in the world, but it ought to work.

This requires some configuration on the live site, to hook the pieces together.