Fix The Incredibly Stupid Cursor Bug
It looks like the ur-cause of the duplicate meetings that have been plaguing us today is the _scheduleCursor, which keeps track of where we are in the Schedule. Which is great in theory, but we were never actually *updating* that cursor in-memory, because I was tired and dumb.

On its own, this was innocuous, because we pull things off the front of the Schedule Queue as we start them. But whenever the Schedule *changes* (which it does fairly often), we would regenerate the whole queue, completely fail to remove the stuff that was already started (because we weren't updating the cursor), and restart all of these already-existing meetings. Hence, bazillions of duplicates.

This fixes that with the one-liner to actually update the cursor where we should have been doing all along.

We also add a belt-and-suspenders check when starting a program item: before doing so, we check the DB to see whether it was already started. In theory this should be irrelevant, but let's be careful.

NB: the reason I didn't notice this before is that bloody Zoom doesn't consider a meeting *real* unless it has two or more people in it. So it has been allowing the system to start lots of conflicting meetings in development without complaining at all.

Finally, this turns the timer back on, to begin starting stuff automatically again, since we believe this fixes the underlying problem.