Fix PII leak for users without Badge Names
The (pretty old) CM-webscrape code defaulted, if you didn't have a Badge Name, to using your username. That seemed to make sense at the time, since the username is the primary key to everything, but it needs to be treated as potential PII, because *that* appears to often be peoples' email address. And that problem becomes Much Worse because we change peoples' Discord nickname to be their Badge Name.

So this just turns off that defaulting, and if the Badge Name is empty, doesn't touch the Discord nickname.