Fix caching when switching from logged-out to logged-in
If you logged in as a program participant, and you had previously looked at the Schedule, and it hadn't changed in Zambia, you would still see the logged-out view without your prep sessions. This fixes that, by distinguishing the logged-out ETag.

(Yes, it's a conceptually ugly fix, but it's easy, safe, and works.)