Fix Profile Page for people who have duplicate login entries
It looks like you can log in with a space before or after your name, and CM silently allows that. (User 96704 apparently did so.) The result is that they have duplicate entries in user_info. That was causing database errors, which in turn (I suspect) means that their Profile Page was broken.

Ideally we would fix the login pathway, but for now the safer approach is to simply tolerate it.