Make links in Art Show & Dealers more usable.
Merged two links together so they will be handled better by screen
readers.

Changed links to be <a> so they would underline and change cursor on
hover and so that they will be handled better by screen readers.

Switched the positioning of <a> and <li> tags so the whole text box
will be a link instead of just the text.