Fix login for members of A'22
Login is failing because we are getting multiple results, and expect to only get one.

The problem is that A'22 (so far) has the same CoC as A'21. So the LEFT JOIN that is pulling in the CoC hits both entries. We need to filter that so that it is only checking your A'21 CoC, not your A'22 one.